+++ ballerina/Ballerina.toml
@@ -1,0 +1,7 @@
+[package]
+org = "ballerina"
+name = "cache"
+version = "@toml.version@"
+
+[[platform.java11.dependency]]
+path = "../native/build/libs/cache-native-@project.version@.jar"
