--- cache-ballerina/Ballerina.toml
@@ -1,10 +1,0 @@
-[package]
-org = "ballerina"
-name = "cache"
-version = "@toml.version@"
-
-[[platform.java11.dependency]]
-path = "../cache-native/build/libs/cache-native-@project.version@.jar"
-groupId = "ballerina"
-artifactId = "cache"
-version = "@project.version@"
