+++ gradle.properties
@@ -1,9 +1,15 @@
 org.gradle.caching=true
-group=org.ballerinalang
+group=io.ballerina.stdlib
 version=2.1.0-beta.3-SNAPSHOT
 ballerinaLangVersion=2.0.0-beta.2.1
 stdlibTaskVersion=2.0.0-beta.2
 stdlibTimeVersion=2.0.0-beta.2
 
 puppycrawlCheckstyleVersion=8.18
-ballerinaGradlePluginVersion=0.9.1
+ballerinaGradlePluginVersion=0.10.0
+testngVersion=6.14.3
+slf4jVersion=1.7.30
+githubSpotbugsVersion=4.0.5
+githubJohnrengelmanShadowVersion=5.2.0
+underCouchDownloadVersion=4.0.4
+researchgateReleaseVersion=2.8.0
